requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
statsmodels>=0.14.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0
kaggle>=1.5.12
pyyaml>=6.0
tqdm>=4.64.0
jupyter>=1.0.0
pytest>=7.2.0